+++ src/refactorings/move-to-existing-file/export-declaration.ts
@@ -1,18 +1,35 @@
 import * as t from "../../ast";
 
 export function getExportDeclaration(
-  path: t.RootNodePath
+  path: t.NodePath
 ): ExportDeclaration | null {
-  return path.isExportNamedDeclaration()
-    ? new NamedExportDeclaration(path)
-    : path.isExportDefaultDeclaration()
-    ? new DefaultExportDeclaration(path)
-    : null;
+  if (t.isRootNodePath(path)) return new NoExportDeclaration(path);
+
+  const { parentPath } = path;
+  if (!parentPath || !t.isRootNodePath(parentPath)) return null;
+
+  return parentPath.isExportNamedDeclaration()
+    ? new NamedExportDeclaration(parentPath)
+    : parentPath.isExportDefaultDeclaration()
+    ? new DefaultExportDeclaration(parentPath)
+    : new NoExportDeclaration(parentPath);
 }
 
 export interface ExportDeclaration {
   readonly parentPath: t.NodePath<t.Program>;
   replaceWith(id: t.Identifier): void;
+}
+
+class NoExportDeclaration implements ExportDeclaration {
+  constructor(private path: t.RootNodePath) {}
+
+  get parentPath(): t.NodePath<t.Program> {
+    return this.path.parentPath;
+  }
+
+  replaceWith() {
+    this.path.remove();
+  }
 }
 
 class NamedExportDeclaration implements ExportDeclaration {