+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Move to existing file now supports moving exported elements. It will preserve the export, so your module API won't break.
+- Move to existing file can now move variable declarations too!
 
 ## [6.11.0] - 2022-02-19 - Dos Oruguitas 🦋